Extract MyApp
