+++ lib/core/ui/my_app.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const MaterialApp(
+      home: Scaffold(
+        body: Center(
+          child: Text('game_saves_backup'),
+        ),
+      ),
+    );
+  }
+}
