+++ lib/main.dart
@@ -1,18 +1,4 @@
 import 'package:flutter/material.dart';
+import 'package:game_saves_backup/core/ui/my_app.dart';
 
 void main() => runApp(const MyApp());
-
-class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const MaterialApp(
-      home: Scaffold(
-        body: Center(
-          child: Text('game_saves_backup'),
-        ),
-      ),
-    );
-  }
-}
